Add tests for new functions (tests for transfers left to do)
